+++ fcs/util/crc32.c
@@ -110,13 +110,13 @@ static const uint32_t crc_lookup[] = { /* CRC polynomial 0xedb88320 */
 0xb40bbe37, 0xc30c8ea1, 0x5a05df1b, 0x2d02ef8d
 };
 
-uint32_t fcs_crc32(const uint8_t *restrict pdata, uint32_t nbytes,
+uint32_t fcs_crc32(const uint8_t *restrict pdata, size_t nbytes,
 uint32_t crc) {
     /* For standard CRC32B results, set crc = 0xffffffffu */
     assert(pdata);
     assert(nbytes);
 
-    uint32_t i;
+    size_t i;
     /* loop over the buffer data */
     #pragma MUST_ITERATE(1)
     for (i = 0; i < nbytes; i++) {
